Fix for Vanilla bug with adding pin after deletion
If you create 2 or more pins and then delete anyone but the last, and
then add a new one then the last will be changed instead.

It is the same if you add a new one but instead of pressing Ok you press
Delete, then the last one will be deleted instead.

I don't know why it happens, in the RequestMapPin function the ID is
correct but when you enter any of the On-functions the ID have changed
to the heighest. This is pressent in vanilla but I got feed up by this
issue so I fixed it for them :wink:

To fix this I store the location as well and fetch the Pin using this
when buttons are pressed.