Don't Reload CityBanner on CQUI_SettingsInitialized
Not sure exactly why but if we don't run Reload on
CQUI_SettingsInitialized for the CityBanner then it seems to work, so I
guess it is run to soon for the objects to be in place or something.

Would be good to test this a bit to make sure that it does not break
anything else instead.